LayoutTests/transitions/color-transition-all.html

<!DOCTYPE>

<html lang="en">
<head>
  <title>Color transition with transition-property: all</title>
  <style type="text/css" media="screen">
    #outer {
      font-size: 200%;
      color: #00F;
      -webkit-transition: all 1s linear;
    }
    
    #inner {
      color: #F00;
    }
    
    #outer.changed {
      color: #0F0;
    }
  </style>
  <script src="transition-test-helpers.js" type="text/javascript" charset="utf-8"></script>
  <script type="text/javascript" charset="utf-8">
    const expectedValues = [
      // [time, element-id, property, expected-value, tolerance]
      [0.5, 'outer', 'color', [0, 127, 127], 2],
      [0.5, 'inner', 'color', [255, 0, 0], 2],
      [0.5, 'inner', '-webkit-text-fill-color', [255, 0, 0], 2],
    ];

    function setupTest()
    {
      document.getElementById('outer').className = 'changed';
    }

    runTransitionTest(expectedValues, setupTest, true);
  </script>
</head>
<body>

  <div id="outer">
    <div id="inner">
      inner
    </div>
    outer
  </div>
  
  <div id="result">
  </div>

</body>
</html>
