LayoutTests/platform/ios-sim/transitions/color-transition-all-expected.txt

CONSOLE MESSAGE: line 240: Failed to pause 'color' transition on element 'outer'
inner
outer
FAIL - "color" property for "outer" element at 0.5s expected: 0,127,127 but saw: 0,255,0
PASS - "color" property for "inner" element at 0.5s saw something close to: 255,0,0
PASS - "-webkit-text-fill-color" property for "inner" element at 0.5s saw something close to: 255,0,0
